Move opcode table to separate file
