+++ DameDis/DameDis.vcxproj
@@ -155,6 +155,7 @@
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">NotUsing</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">NotUsing</PrecompiledHeader>
     </ClCompile>
+    <ClCompile Include="OpcodeTables.h" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Cartridge.h" />
