+++ DameDis/Dissassembler.h
@@ -3,16 +3,6 @@
 #include <vector>
 #include <string>
 
-struct Opcode {
-	std::string mnemonic;
-	uint8_t operand_bytes;
-};
-
-struct CB_Opcode : public Opcode {
-	CB_Opcode(std::string mnemonic) : Opcode{ mnemonic, 0 } {};
-	std::string mnemonic;
-};
-
 constexpr uint16_t NO_OPERAND = 0x10000;
 
 struct Instruction {
@@ -35,8 +25,6 @@ public:
 	std::string GetMnemonic(uint16_t index);
 	std::string GetOperand(uint16_t index);
 
-	static Opcode opcodeTable[256];
-	static CB_Opcode cb_opcodeTable[256];
 private:
 	uint8_t fetch();
 	void StoreNextInstruction();
